move match ancestry function to shared file
