+++ lib/checks/navigation/heading-order-after.js
@@ -1,8 +1,10 @@
+import { matchAncestry } from '../../core/utils';
+
 export default function headingOrderAfter(results) {
   // Construct a map of all headings on the page
   const headingOrder = getHeadingOrder(results);
   results.forEach(result => {
-    result.result = getHeadingOrderOutcome(result, headingOrder)
+    result.result = getHeadingOrderOutcome(result, headingOrder);
   });
   return results;
 }
@@ -11,20 +13,20 @@ export default function headingOrderAfter(results) {
  * Determine check outcome, based on the position of the result in the headingOrder
  */
 function getHeadingOrderOutcome(result, headingOrder) {
-  const index = findHeadingOrderIndex(headingOrder, result.node.ancestry)
+  const index = findHeadingOrderIndex(headingOrder, result.node.ancestry);
   const currLevel = headingOrder[index]?.level ?? -1;
   const prevLevel = headingOrder[index - 1]?.level ?? -1;
 
   // First heading always passes
   if (index === 0) {
-    return true
-  };
+    return true;
+  }
   // Heading not in the map
   if (currLevel === -1) {
-    return undefined; 
+    return undefined;
   }
   // Check if a heading is skipped
-  return (currLevel - prevLevel <= 1)
+  return currLevel - prevLevel <= 1;
 }
 
 /**
@@ -73,7 +75,7 @@ function mergeHeadingOrder(mergedHeadingOrder, result) {
 
 /**
  * Determine where the iframe results fit into the top-level heading order
- * 
+ *
  * If a frame has no headings, but it does have iframes we might not have a result.
  * We can account for this by finding the closest ancestor we do know about.
  */
@@ -83,7 +85,7 @@ function getFrameIndex(headingOrder, frameAncestry) {
     if (index !== -1) {
       return index;
     }
-    frameAncestry = shortenArray(frameAncestry, 1)
+    frameAncestry = shortenArray(frameAncestry, 1);
   }
   return -1;
 }
@@ -103,25 +105,6 @@ function findHeadingOrderIndex(headingOrder, ancestry) {
 function addFrameToHeadingAncestry(heading, frameAncestry) {
   const ancestry = frameAncestry.concat(heading.ancestry);
   return { ...heading, ancestry };
-}
-
-/**
- * Check if two ancestries are identical
- */
-function matchAncestry(ancestryA, ancestryB) {
-  if (ancestryA.length !== ancestryB.length) {
-    return false;
-  }
-  return ancestryA.every((selectorA, index) => {
-    const selectorB = ancestryB[index];
-    if (!Array.isArray(selectorA)) {
-      return selectorA === selectorB;
-    }
-    if (selectorA.length !== selectorB.length) {
-      return false;
-    }
-    return selectorA.every((str, index) => selectorB[index] === str);
-  });
 }
 
 /**
