+++ lib/core/utils/index.js
@@ -53,6 +53,7 @@ export {
   matchesExpression,
   convertSelector
 } from './matches';
+export { default as matchAncestry } from './match-ancestry';
 export { default as memoize } from './memoize';
 export { default as mergeResults } from './merge-results';
 export { default as nodeSorter } from './node-sorter';
